Address FreeTDS 0.95 and ODBC compiler warnings.
* Harmonize type usage, via explicit casts where necessary (and safe).
  In particular, ensure that upstream tests' tds_thread_proc routines
  know to return DWORD rather than void* on Windows.
* Ensure that any headers checking (TDS_)HAVE_* #include <config.h> first.
* Keep static assertions warning-free by using _Static_assert where
  available, and the standard offsetof macro from <stddef.h>.
* Ensure that members of types that should be unhidden have unhidden
  forward declarations; likewise for their members, recursively.
* For LLVM Clang, substitute __attribute__((visibility("default")))
  for __attribute__((externally_visible)), which it doesn't support.
* When including wspiapi.h, temporarily suppress warning C4191
  ("unsafe conversion", mostly to FARPROC), which adds a lot of noise.
* ctlib.h: #define _CS_CURS_TYPE_* to TDS_CURSOR_STATE_* rather than
  (typelessly) duplicating their values.
* dblib/bcp.c (_bcp_get_col_data): Remove a redundant test for non-negativity.
* dblib/dblib.c: Express -2?\194?\179?\194?\185 in a manner all compilers accept happily.
* odbc/unittests/attributes.c: Tell the compiler fatal() will never return.
* odbc/unittests/common.h: When the ODBC 3 API is available, favor
  SQLSetStmtAttr over SQLSetStmtOption, which is deprecated.
* odbc/unittests/tables.c: On Windows, alias strcasecmp to _stricmp, as
  the underscoreless version is deprecated.
* replacements/tds_cond.c, win32/winsetup.c: Explicitly use
  narrow-character-based APIs as appropriate.
* tds/convert.c: Omit tds_hex_digits's length specification, which
  was unnecessary and didn't account for the implicit trailing NUL.
* tds/convert.c, tds/unittests/t0007.c: Tweak macro-wrapped
  function names to avoid confusing IDEs.
* tds/gssapi.c: Suppress GSSAPI deprecation warnings on Mac OS X.
* tds/md[45].c: Use TDS_PUT_(U)A4 to populate buf cleanly.
* tds/query.c: #include <sys/socket.h> if available, for send().
* tds/sspi.c: Comment out code that used a deprecated function to check
  for ancient Windows versions, on which its modern substitute isn't
  available in the first place.
* thread.h: Correct tds_thread_proc and TDS_THREAD_PROC_DECLARE; Windows
  threads are expected to return DWORD, not void*.

JIRA: CXX-4155.
TRUNK-0.64: 190053, 355781, 356114.  AUTHOR: ivanovp.
TRUNK-0.64: 442324, 487448.  AUTHOR: ucko.

git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@70343 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ include/dbapi/driver/ftds95/freetds/des.h
@@ -16,8 +16,8 @@ typedef struct des_key
 } DES_KEY;
 
 void tds_des_set_odd_parity(des_cblock key);
-int tds_des_ecb_encrypt(const void *plaintext, int len, DES_KEY * akey, des_cblock output);
-int tds_des_set_key(DES_KEY * dkey, const des_cblock user_key, int len);
+int tds_des_ecb_encrypt(const void *plaintext, size_t len, DES_KEY * akey, des_cblock output);
+int tds_des_set_key(DES_KEY * dkey, const des_cblock user_key, size_t len);
 void tds_des_encrypt(DES_KEY * key, des_cblock block);
 void _mcrypt_decrypt(DES_KEY * key, unsigned char *block);
