+++ src/dbapi/driver/ftds95/freetds/replacements/iconv.c
@@ -99,14 +99,14 @@ get_utf8(const unsigned char *p, size_t len, ICONV_CHAR *out)
 	while(--l)
 		uc = (uc << 6) | (*p++ & 0x3f);
 	*out = uc;
-	return len;
+    return (int)len;
 }
 
 static int
 put_utf8(unsigned char *buf, size_t buf_len, ICONV_CHAR c)
 {
 #define MASK(n) ((0xffffffffu << (n)) & 0xffffffffu)
-	size_t o_len;
+    int o_len;
 	unsigned mask;
 
 	if ((c & MASK(7)) == 0) {
